requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bokeh==3.4.1
branca==0.7.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
dataclasses-json==0.6.4
deepgram-sdk==3.2.4
distro==1.9.0
engineering-notation==0.10.0
faiss-cpu
Flask
folium==0.16.0
frozenlist==1.4.1
google==3.0.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
langchain==0.2.9  # Specify a compatible version for langchain
langchain_community==0.2.9  # Specify a compatible version for langchain_community
langchain_openai==0.1.15  # Specify a compatible version for langchain_openai
langchain-openai
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.32.0  # Use a version compatible with langchain_openai
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.21.0
proto-plus==1.23.0
protobuf==4.25.3
psutil
pydantic==2.6.4
pydantic_core==2.16.3
PyPDF2
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tenacity==8.2.3
tk==0.1.0
tk-tools==0.16.0
tornado==6.4
tqdm==4.66.2
twilio
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
verboselogs==1.7
websockets==12.0
xyzservices==2024.4.0
yarl==1.9.4
